+++ apps-intern/ungdomsytelse-veileder/storybook/mock-data/soker1/søkerMock.ts
@@ -1,8 +1,0 @@
-export const søkerMock = {
-    aktørId: '2320509955297',
-    fødselsdato: '2005-06-02',
-    fødselsnummer: '02869599258',
-    fornavn: 'Test',
-    mellomnavn: null,
-    etternavn: 'Brukeresen',
-};