+++ apps-intern/ungdomsytelse-veileder/vite.config.ts
@@ -24,6 +24,11 @@ export default defineConfig({
         },
         port: 8088,
     },
+    define: {
+        __IS_VEILEDER_DEMO__: false,
+        __IS_GITHUB_PAGES__: false,
+        __VIS_DEMO_BRUKERE__: false,
+    },
     resolve: {},
     build: {
         sourcemap: true,