+++ apps-intern/ungdomsytelse-veileder/vite.dev.config.ts
@@ -67,6 +67,11 @@ export default defineConfig({
             },
         },
     },
+    define: {
+        __IS_VEILEDER_DEMO__: true,
+        __IS_GITHUB_PAGES__: false,
+        __VIS_DEMO_BRUKERE__: true,
+    },
     base: '/sif-brukerdialog/ungdomsytelse-veileder/',
     build: {
         sourcemap: true,