+++ apps/ekstra-omsorgsdager-andre-forelder-ikke-tilsyn/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @navikt/ekstra-omsorgsdager-andre-forelder-ikke-tilsyn
 
+## 1.35.71
+
+### Patch Changes
+
+- ca64a56: axios 1.3.5 fix
+
+## 1.35.70
+
+### Patch Changes
+
+- Pakkeoppdatering - @isaacs/brace-expansion5.0.1
+
+## 1.35.69
+
+### Patch Changes
+
+- Updated dependencies
+    - @navikt/sif-common-hooks@0.3.10
+    - @navikt/sif-common-soknad-ds@26.1.37
+
 ## 1.35.68
 
 ### Patch Changes