+++ apps/omsorgsdager-kalkulator/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @navikt/omsorgsdager-kalkulator
 
+## 3.14.3
+
+### Patch Changes
+
+- ca64a56: axios 1.3.5 fix
+
+## 3.14.2
+
+### Patch Changes
+
+- Pakkeoppdatering - @isaacs/brace-expansion5.0.1
+
 ## 3.14.1
 
 ### Patch Changes