+++ apps/pleiepenger-sykt-barn/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @navikt/pleiepenger-sykt-barn
 
+## 4.11.66
+
+### Patch Changes
+
+- ca64a56: axios 1.3.5 fix
+
+## 4.11.65
+
+### Patch Changes
+
+- Pakkeoppdatering - @isaacs/brace-expansion5.0.1
+
+## 4.11.64
+
+### Patch Changes
+
+- Updated dependencies
+    - @navikt/sif-common-hooks@0.3.10
+    - @navikt/sif-common-soknad-ds@26.1.37
+
 ## 4.11.63
 
 ### Patch Changes