+++ packages/k9-sak-innsyn-api/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @navikt/k9-brukerdialog-prosessering-api
 
+## 0.2.12
+
+### Patch Changes
+
+- Inntektsmelding og splitt av sak i Dine pleiepenger. Tilhørende codegen + noe felleskoderoppdateringer
+
 ## 0.2.11
 
 ### Patch Changes