+++ server-ungdomsytelse-veileder/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @navikt/sif-server
 
+## 0.0.56
+
+### Patch Changes
+
+- Pakkeoppdatering - @isaacs/brace-expansion5.0.1
+
 ## 0.0.55
 
 ### Patch Changes