+++ server/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @navikt/sif-server
 
+## 0.4.57
+
+### Patch Changes
+
+- Pakkeoppdatering - @isaacs/brace-expansion5.0.1
+
 ## 0.4.56
 
 ### Patch Changes